refactor: 코드 리뷰 전 최종 점검
* chore: import문 정리

* fix: 스타일 보완

* chore: 요구사항 최신화
