+++ docs/요구사항.md
@@ -8,7 +8,7 @@
   - Endpoint만 변경하면 언제든 Real API를 바라볼 수 있다고 가정하고 상상한다.
   - Real API 없이 로컬에서만 동작하는 상태로 리뷰 받는 것이 기본 원칙
 - [x] 상태 관리 라이브러리를 선택하고 적용한다. (react-query)
-  - [ ] 전략을 세우고 PR 본문에 내용을 작성한다.
+  - [x] 전략을 세우고 PR 본문에 내용을 작성한다.
   - [ ] 선택한 상태 관리 라이브러리에 대응되는 테스트 전략을 세우고 PR 본문에 내용을 작성한다.
   - [ ] 없다면 React Testing Library & Jest를 활용해 자유로운 단위의 테스트라도 진행한다.
