+++ src/components/index.ts
@@ -1,6 +1,15 @@
 export * from "./Header";
+export * from "./SectionHeader";
+export * from "./EstimatedPaymentSide";
 
 export * from "./CartItem";
+export * from "./CartItems";
+export * from "./CartItemsSkeleton";
+
 export * from "./OrderItem";
 export * from "./ProductItem";
 export * from "./PrimaryButton";
+
+export * from "./PrimaryButton";
+export * from "./Spinner";
+export * from "./UnknownError";
