+++ src/css/page/order.css
@@ -1,5 +1,8 @@
 .order-section {
   padding: 24px 300px;
+
+  margin: 0 auto;
+  max-width: 1920px;
 }
 
 .order-section__title {
