+++ src/css/page/product.css
@@ -4,6 +4,7 @@
   align-items: center;
   flex-direction: column;
 
+  margin: 0 auto;
   max-width: 1920px;
 }
